reports_parser/pyproject.toml
[tool.poetry]
name = "reports_parser"
version = "0.1.0"
description = ""
authors = ["AidarRakhmatullin <nedopolz@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.10"
loguru = "^0.5.3"
aiohttp = "^3.7.4"
SQLAlchemy = "^1.4.11"
python-decouple = "^3.4"
asyncpg = "^0.24.0"
openpyxl = "^3.0.7"
pandas = "^1.3.2"
gspread = "^4.0.1"
beautifulsoup4 = "^4.9.3"
APScheduler = "^3.7.0"
asgiref = "^3.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
